refactor secondary template resolver to its own class
