+++ service/src/main/java/bio/terra/pipelines/App.java
@@ -10,8 +10,6 @@ import org.springframework.orm.jpa.JpaTransactionManager;
 import org.springframework.retry.annotation.EnableRetry;
 import org.springframework.transaction.PlatformTransactionManager;
 import org.springframework.transaction.annotation.EnableTransactionManagement;
-import org.thymeleaf.templatemode.TemplateMode;
-import org.thymeleaf.templateresolver.ClassLoaderTemplateResolver;
 
 @SpringBootApplication()
 @ComponentScan(
@@ -46,22 +44,5 @@ public class App {
   @Bean("transactionManager")
   public PlatformTransactionManager getTransactionManager() {
     return new JpaTransactionManager();
-  }
-
-  /**
-   * This bean is used to resolve the location of the Thymeleaf templates that are used to generate
-   * the OpenAPI documentation. The default resolver is used to resolve the location of the Swagger
-   * UI index.html file in templates/.
-   */
-  public ClassLoaderTemplateResolver secondaryTemplateResolver() {
-    ClassLoaderTemplateResolver secondaryTemplateResolver = new ClassLoaderTemplateResolver();
-    secondaryTemplateResolver.setPrefix("static/");
-    secondaryTemplateResolver.setSuffix(".yml");
-    secondaryTemplateResolver.setTemplateMode(TemplateMode.TEXT);
-    secondaryTemplateResolver.setCharacterEncoding("UTF-8");
-    secondaryTemplateResolver.setOrder(1);
-    secondaryTemplateResolver.setCheckExistence(true);
-
-    return secondaryTemplateResolver;
   }
 }
